refactor: entry points and modularize page scripts
Split main.ts into per-page entry scripts under src/pages, moving shared logic to src/utils/common.ts. Updated CSS imports and resource paths, consolidated base styles into main.css, and adjusted type definitions for window extensions. Updated templates to load new per-page assets and scripts. This improves maintainability and enables more granular page-specific logic and styling.
