+++ settings.yaml
@@ -87,8 +87,8 @@ spec:
           name: css_resource_location
           if: "$is_custom_resource_locations === true"
           label: CSS 资源位置
-          value: "/themes/howiehz-higan/assets/dist/style.css"
-          help: "指定 CSS 资源的位置，例如：https://cdn.example.com/style.css"
+          value: "/themes/howiehz-higan/assets/dist/common.css"
+          help: "指定 CSS 资源的位置，例如：https://cdn.example.com/common.css"
         - $formkit: text
           name: instantpage_resource_location
           if: "$is_custom_resource_locations === true"
