+++ src/pages/archives.ts
@@ -1,0 +1,2 @@
+import "../utils/common";
+import "../utils/animation";
