+++ src/pages/author.ts
@@ -1,0 +1,2 @@
+import "../utils/common";
+import "../styles/pages/author.css";
