+++ src/pages/links.ts
@@ -1,0 +1,2 @@
+import "../utils/common";
+import "../styles/pages/links.css";
