+++ src/pages/page.ts
@@ -1,0 +1,3 @@
+import "../utils/common";
+import "../utils/animation";
+import "../utils/create-toc";
