+++ src/pages/photos.ts
@@ -1,0 +1,2 @@
+import "../utils/common";
+import "../styles/pages/photos.css";
