+++ src/utils/animation.ts
@@ -1,5 +1,6 @@
 // 导入类型定义（仅类型导入，不会在运行时包含）
-import type {} from "../types/window";
+import type {} from "../types/window-animation";
+import "../styles/mixins/animation.css";
 
 // 使此文件成为模块
 export {};
