+++ templates/archives.html
@@ -1,8 +1,22 @@
 <!doctype html>
 <th:block
   xmlns:th="https://www.thymeleaf.org"
-  th:replace="~{fragments/layout :: html(title=#{page.archives.page-title}, content = ~{::content}, header = ~{fragments/header::content}, navbar = ~{})}"
+  th:replace="~{fragments/layout :: html(
+    title = #{page.archives.page-title},
+    head = ~{::head},
+    content = ~{::content},
+    header = ~{fragments/header::content},
+    navbar = ~{}
+  )}"
 >
+  <th:block th:fragment="head">
+    <link
+      rel="stylesheet"
+      th:href="@{/assets/dist/animation.css(v=${theme.spec.version})}"
+      href="assets/dist/animation.css"
+    />
+    <script th:src="@{/assets/dist/archives.js(v=${theme.spec.version})}" type="module" defer></script>
+  </th:block>
   <th:block th:fragment="content">
     <div id="archive" class="content flex-auto">
       <div class="flex-1">
