+++ templates/author.html
@@ -1,8 +1,22 @@
 <!doctype html>
 <th:block
   xmlns:th="https://www.thymeleaf.org"
-  th:replace="~{fragments/layout :: html(title=${site.title}, content = ~{::content}, header = ~{fragments/header::content}, navbar = ~{})}"
+  th:replace="~{fragments/layout :: html(
+    title = ${site.title},
+    head = ~{::head},
+    content = ~{::content},
+    header = ~{fragments/header::content},
+    navbar = ~{}
+  )}"
 >
+  <th:block th:fragment="head">
+    <link
+      rel="stylesheet"
+      th:href="@{/assets/dist/author.css(v=${theme.spec.version})}"
+      href="assets/dist/author.css"
+    />
+    <script th:src="@{/assets/dist/author.js(v=${theme.spec.version})}" type="module" defer></script>
+  </th:block>
   <th:block th:fragment="content">
     <article class="author" itemscope itemtype="http://schema.org/BlogPosting">
       <section class="about">
