+++ vite.config.ts
@@ -5,6 +5,7 @@ import { defineConfig } from "vite";
 import PurgeIcons from "vite-plugin-purge-icons";
 
 export default defineConfig({
+  base: "/themes/howiehz-higan/assets/dist/",
   plugins: [
     PurgeIcons({
       content: ["./templates/*.html"],
@@ -18,13 +19,18 @@ export default defineConfig({
     rollupOptions: {
       input: {
         main: path.resolve(__dirname, "src/main.ts"),
+        index: path.resolve(__dirname, "src/pages/index.ts"),
+        post: path.resolve(__dirname, "src/pages/post.ts"),
+        page: path.resolve(__dirname, "src/pages/page.ts"),
+        moment: path.resolve(__dirname, "src/pages/moment.ts"),
+        archives: path.resolve(__dirname, "src/pages/archives.ts"),
+        photos: path.resolve(__dirname, "src/pages/photos.ts"),
+        links: path.resolve(__dirname, "src/pages/links.ts"),
+        author: path.resolve(__dirname, "src/pages/author.ts"),
       },
       output: {
-        entryFileNames: "main.js",
-        assetFileNames: (assetInfo) => {
-          if (assetInfo.names && assetInfo.names.includes("main.css")) {
-            return "style.css";
-          }
+        entryFileNames: "[name].js",
+        assetFileNames: () => {
           return "[name][extname]";
         },
       },
